update golang version for Docker and switch to CMD
with EXEC syntax.

This change was actually done because of the problem that when using
ENTRYPOINT with shell syntax, in the Docker container first a shell
command is launched which in turn executes the go command afterwards.

The problem with that is when stopping the container the shell command
would receive the SIGTERM signal and it does not pass it somehow to the
running go-process. This means the container does not exit after
receiving the SIGTERM but docker will invoke, usually after 10 seconds,
a SIGKILL. This is somewhat unclean.